Fix namespace parsing issue on module install
